chore(test): move ProductAPI to `api` package, DummyRequest to `data` package
